+++ src/main/java/by/it_academy/jd2/MJD29522/dao/orm/SingerDaoHibernate.java
@@ -2,10 +2,9 @@ package by.it_academy.jd2.MJD29522.dao.orm;
 
 import by.it_academy.jd2.MJD29522.dao.api.ISingerDao;
 import by.it_academy.jd2.MJD29522.dao.orm.api.IManager;
+import by.it_academy.jd2.MJD29522.dto.SingerDTO;
 import by.it_academy.jd2.MJD29522.entity.SingerEntity;
-
 import javax.persistence.EntityManager;
-import java.util.ArrayList;
 import java.util.List;
 
 public class SingerDaoHibernate implements ISingerDao {
@@ -18,19 +17,18 @@ public class SingerDaoHibernate implements ISingerDao {
 
     @Override
     public List<SingerEntity> get() {
-        List<SingerEntity> singers = new ArrayList<>();
         EntityManager entityManager = null;
         try {
             entityManager = manager.getEntityManager();
             entityManager.getTransaction().begin();
             List<SingerEntity>singerEntityList = entityManager.createQuery
                     ("from SingerEntity ORDER BY id", SingerEntity.class).getResultList();
-            entityManager.getTransaction().commit();
-            singers.addAll(singerEntityList);
-            return singers;
 
+            entityManager.getTransaction().commit();
+
+            return singerEntityList;
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
@@ -42,16 +40,16 @@ public class SingerDaoHibernate implements ISingerDao {
     }
 
     @Override
-    public boolean add(String newSinger) {
+    public boolean add(SingerDTO newSinger) {
         EntityManager entityManager = null;
         try {
             entityManager = manager.getEntityManager();
             entityManager.getTransaction().begin();
-            entityManager.persist(new SingerEntity(newSinger));
+            entityManager.persist(new SingerEntity(newSinger.getName()));
             entityManager.getTransaction().commit();
             return true;
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
@@ -69,12 +67,12 @@ public class SingerDaoHibernate implements ISingerDao {
             entityManager = manager.getEntityManager();
             entityManager.getTransaction().begin();
             SingerEntity singerToUpdate = entityManager.find(SingerEntity.class, singerEntity.getId());
-            if(singerToUpdate != null){
+            if(singerToUpdate != null && singerEntity.getVersion().equals(singerToUpdate.getVersion())){
                 entityManager.merge(singerEntity);
             }
             entityManager.getTransaction().commit();
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
@@ -86,19 +84,19 @@ public class SingerDaoHibernate implements ISingerDao {
     }
 
     @Override
-    public boolean delete(long id) {
+    public boolean delete(long id, long version) {
         EntityManager entityManager = null;
         try {
             entityManager = manager.getEntityManager();
             entityManager.getTransaction().begin();
             SingerEntity singerToRemoved = entityManager.find(SingerEntity.class, id);
-            if(singerToRemoved != null){
+            if(singerToRemoved != null && singerToRemoved.getVersion().equals(version)){
                 entityManager.remove(singerToRemoved);
             }
             entityManager.getTransaction().commit();
             return singerToRemoved != null;
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
@@ -110,16 +108,16 @@ public class SingerDaoHibernate implements ISingerDao {
     }
 
     @Override
-    public boolean exist(long id) {
+    public SingerEntity exist(long id) {
         EntityManager entityManager = null;
         try {
             entityManager = manager.getEntityManager();
             entityManager.getTransaction().begin();
             SingerEntity singer = entityManager.find(SingerEntity.class, id);
             entityManager.getTransaction().commit();
-           return singer != null;
+           return singer;
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
@@ -144,13 +142,35 @@ public class SingerDaoHibernate implements ISingerDao {
             }
             return singer.getName();
         } catch (Exception e) {
-            if(entityManager != null){
+            if(entityManager != null && entityManager.getTransaction().isActive()){
                 entityManager.getTransaction().rollback();
             }
             throw new RuntimeException("Ошибка в базе данных", e);
         }
         finally {
             if (entityManager != null) {
+                entityManager.close();
+            }
+        }
+    }
+    @Override
+    public SingerEntity getCard(long id){
+        EntityManager entityManager = null;
+        try{
+            entityManager = manager.getEntityManager();
+            entityManager.getTransaction().begin();
+            SingerEntity singerEntity = entityManager.find(SingerEntity.class, id);
+            entityManager.getTransaction().commit();
+            if(singerEntity != null){
+                return singerEntity;
+            }else throw new IllegalArgumentException("Артиста с id " + id + " не найдено");
+        } catch (Exception e) {
+            if(entityManager != null && entityManager.getTransaction().isActive()){
+                entityManager.getTransaction().rollback();
+            }
+            throw new RuntimeException("Ошибка в базе данных", e);
+        }finally {
+            if(entityManager != null){
                 entityManager.close();
             }
         }
